Added delete functionality for each asset or liability, as well as networth and asset and liability totals.

+++ app/models/asset.go
@@ -8,5 +8,5 @@ type Asset struct {
   gorm.Model 
   Name string `gorm:"size:64"`
   Type string `gorm:"size:16"`
-  Balance float32
+  Balance float64
 }
